refactoring) Rank enum 외부로 분리
