+++ src/main/java/lotto/view/OutputView.java
@@ -1,6 +1,7 @@
 package lotto.view;
 
 import lotto.model.Lotto;
+import lotto.model.Rank;
 
 public class OutputView {
     /*
@@ -27,7 +28,7 @@ public class OutputView {
         System.out.println("당첨 통계");
         System.out.println("---------");
 
-        for (Lotto.Rank rank : Lotto.Rank.values()) {
+        for (Rank rank : Rank.values()) {
             System.out.printf("%d개 일치 (%d원)- %d개\n", rank.getMatchNumber(), rank.getPrize(), rank.getCount());
         }
     }
